fix(ci): add webpack build gate and route-export-safe helpers
