+++ app/api/stripe/webhook/route.ts
@@ -14,6 +14,7 @@ import {
   type CatalogProduct,
 } from "@/lib/commerce/catalog";
 import { trackAnalyticsEvent } from "@/lib/analytics/events";
+import { getDiscountRedeemedPayload } from "@/lib/stripe/webhook-discount";
 import { createStripeClient, getStripeWebhookSecret } from "@/lib/stripe/server";
 
 const SUPPORTED_EVENTS = new Set([
@@ -109,27 +110,6 @@ async function fulfillCheckoutSession(stripe: Stripe, session: Stripe.Checkout.S
     productId: product.id,
     userId,
   });
-}
-
-export type DiscountRedeemedPayload = {
-  sessionId: string;
-  productId: string | null;
-  amountDiscount: number;
-  currency: string | null;
-};
-
-export function getDiscountRedeemedPayload(
-  session: Stripe.Checkout.Session
-): DiscountRedeemedPayload | null {
-  const amountDiscount = session.total_details?.amount_discount ?? 0;
-  if (amountDiscount <= 0) return null;
-
-  return {
-    sessionId: session.id,
-    productId: session.metadata?.fs_product_id ?? null,
-    amountDiscount,
-    currency: session.currency ?? null,
-  };
 }
 
 export async function POST(request: Request) {
